+++ src/Calendar/constants.ts
@@ -1,0 +1,32 @@
+/* ======================================================================== *
+ * Copyright 2026 HCL America Inc.                                          *
+ * Licensed under the Apache License, Version 2.0 (the "License");          *
+ * you may not use this file except in compliance with the License.         *
+ * You may obtain a copy of the License at                                  *
+ *                                                                          *
+ * http://www.apache.org/licenses/LICENSE-2.0                               *
+ *                                                                          *
+ * Unless required by applicable law or agreed to in writing, software      *
+ * distributed under the License is distributed on an "AS IS" BASIS,        *
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
+ * See the License for the specific language governing permissions and      *
+ * limitations under the License.                                           *
+ * ======================================================================== */
+
+/**
+ * Date format constants used throughout the Calendar component
+ */
+export const DATE_KEY_FORMAT = 'YYYY-MM-DD';
+export const MONTH_YEAR_FORMAT = 'MMMM YYYY';
+export const FULL_DATE_FORMAT = 'MMMM D, YYYY';
+export const DAY_FORMAT = 'D';
+export const DAY_ABBREVIATION_FORMAT = 'ddd';
+export const WEEK_RANGE_FORMAT = 'MMM D';
+export const WEEK_RANGE_END_FORMAT = 'MMM D, YYYY';
+
+/**
+ * Default calendar dimensions
+ */
+export const DEFAULT_CALENDAR_WIDTH = '841px';
+export const DEFAULT_CALENDAR_HEIGHT = '805px';
+export const DEFAULT_MIN_HEIGHT = '400px';