+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added the `endAdornmentIconButton` prop to the `TextField` component to allow for an icon button at the end of the input field, and rearranged the order of the icon buttons.
+- Added the `Calendar` component.
 
 ### Fixed
 - Fixed tooltip cropping in ProgressBar when asset names exceed container width.